Target .NET 10 and enable nullable refs
Upgrade the library and test app to .NET 10 while keeping .NET Framework 4.8 support, update package metadata and dependencies, and replace the legacy solution file with a .slnx. This also adds Windows platform annotations, enables nullable reference types across the hardware models, initializes collections inline, and cleans up network/MAC address handling and related helpers. README and the sample app were updated to reflect the new requirements and non-interactive console behavior.

+++ HardwareIds.NET/HardwareIds_BaseBoard.cs
@@ -17,7 +17,7 @@
                 {
                     InHwid.Baseboards.Add(new HwBaseboard
                     {
-                        Id = (int) InHwid.Baseboards.Count,
+                        Id = InHwid.Baseboards.Count,
                         Manufacturer = Baseboard.Manufacturer,
                         Model = Baseboard.Product,
                         Version = Baseboard.Version,
